Server may return 'Bad Machine'. Handle this in the new client
An issue appeared when an illegal machine id was passed, causing a raw text response from the server which fails to be converted into a dictionary later on. This is now handled by returning to the main loop, and resetting the machine id. This should never happen I had thought, but seems it can. One possible reason was failed FileIO, so now we only update the machine.txt when the ID has actually changed.

Resolves #43