Rewrite the worker test selection scheme
Originally, we would have a TP value for each test. That value would determine how many games are played per workload. The server would then assign tests with equal probability of the same prio.

Now, all workload sizes are static (based on thread counts of the test and the machine), and TP is used to weight the resources allocated to a test. The server fill ensure 'fair' distribution of resources within a window of error, and place a priority on assigning machines with the same task they already had