Explicitly pkill engine processes
OpenBench spawns cutechess-ob(.exe), which can be killed.
Engines are named Engine-GITSHA(-NETSHA), and can be killed.

In theory, if you ran OpenBench at the same time as some other application, you could kill the wrong thing. But you would be ill advised to run an OpenBench Client while doing regular work on a machine. Likewise there is little to no reason to name your cutechess and engines in the same way as OpenBench.

Resolves https://github.com/AndyGrant/OpenBench/issues/162