Critical bug fix when setting test.errors
Somehow in 10 years I've not yet encounted that in python the expression 3 or 3 evaluates to 3, and not True. This allowed for an attempt to save non-boolean values into the test.error field. Which produced seemingly infinite time losses on some tests I was running with some of the first versions of Torch to ever be on OpenBench.

Also, don't report time loss PGNS back. No one cares about them, and flagging the tests is annoying.