Bugfix cross-engine play test mode selection
If you attempt to set a test mode for the dev engine, which the base engine does not have, we produce errors. The only reason we need to check the base engine's understanding of the test mode is to set Base Options. We can fall back to the STC settings, since we already copy the Threads= and Hash= portion from the Dev options