Add balance_engine_throughputs as a configuration option
Load balances tests with relation to each engine. If set to true, the effective throughput during work allocation will be scaled linearly with the number of tests which share the same dev engine.

Namely, it can be used in a large framework to ensure equal allocation of resources across various engines. This comit will be followed by one which allows the user to specify an engine to the client, via --engine MyEngine, to force their worker to selfishly select their engine.

We will account for cases where the selfish-engine is not a top-prio workload, and not play the low prio one.
We will account for the other workers, by ignoring this worker's contribution to overall worker distrubtion. This ensures that it is a strict net gain to attach your own machines to a framework