Improve worker assignment algorithm
If a handful of tests are present, its possible for one test to get multiple times more resources than is fair, without making any other tests be considered unbalacned. IE, the one test steals a small amount of resources from all other tests.

Fix this by bounding the definition of balance as +- 25%, instead of having a one-sided bound.

Also, when computing ratios, compute them as if the machine we are assigning has been assigned to each test. Ie, the ratios IF the worker is assigned to the specific test. Likewise, the fair ratio should account for this machine's threads.