Various fixes for Windows ( and maybe Linux )
using Make to tell the compiler to produce a binary to the /Engines/ dir fails on Windows when the tmpfs is not on the same disk as the Client. This might have also occured on Windows. shutil resolves that.

Take extra care to always maintain the possible .exe extension on files when doing this. This is done twice in the shutil line; first making sure to pickup the true file output name, and secondly to only use the root of the out_path and not the basename