Fixes for relative vs absolute file paths when building (#193)
The output binary is saved alongside the makefile initially, and moved after the fact. This was done to account for Windows and cross-drive issues due to temp filesystems. Because we are saving alongside the makefile, we can just use a relative path, ie the basename of the engine binary, to avoid possible issues in some engine build scripts.

Also, check process exitcode instead of just checking for the binary when building
Also, pretty-print the compiler error, even though it is sent to the server