Force Windows binaries to have .exe extensions
Technically, authors really should have their makefiles attached the .exe extension on Windows. But they seldom do. Its easier to add this hack fix, than it is to enforce that standard, and help people debug.

Worth noting that this is only an issue for Windows, in shells that don't let you execute files, IE powershell/cmd.