Attempt to resolve cascading reimports on worker updates (#265)
TLDR: reloading worker is not sufficient, as the modules loaded by worker will not also be reloaded. Resolving this is non-trivial, mostly due to the nature of from x import ... statements, which would require explicit reloading of each loaded entity. Swapping the entire client to use import x means it should be possible to reload just the list of python modules, at the top of the worker.